refactor handler request methods

+++ systemTests/testScenarios/when/when_steps.py
@@ -12,43 +12,44 @@ class WhenSteps:
         print(self.__class__)
 
     def create_bucket_called_for_simple_bucket(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/createbucket/'
+        handler = 'createbucket'
         body = {'bucket': 'systemtest-textbucket'}
-        response = requests.post(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def copy_dir_called_for_whole_directory(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/copydir/'
+        handler = 'copydir'
         body = {'bucket': self._context.standard_bucket_name}
-        response = requests.post(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def copy_dir_called_for_sub_directory(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/copydir/'
+        handler = 'copydir'
         body = {'bucket': 'systemtest-textbucket', 'directory': 'sub_dir_A'}
-        response = requests.post(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def list_present_called_for_simple_bucket_whole_directory_structure(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/listpresent/'
+        handler = 'listpresent'
         body = {'bucket': 'systemtest-textbucket'}
-        response = requests.post(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def copy_new_called_for_whole_directory(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/copynew/'
+        handler = 'copynew'
         body = {'bucket': 'systemtest-textbucket'}
-        response = requests.post(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def delete_file_is_called_for_simple_file_in_s3(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/deletefile/'
+        handler = 'deletefile'
         body = {'bucket': 'systemtest-textbucket', 'file': 'testA.txt'}
-        response = requests.delete(url, json=body)
-        self._context.http_client_response = response
+        self.make_request(handler, body)
 
     def copy_update_called_for_whole_directory(self):
-        url = f'http://{self._context.host_address}:{self._context.app_port}/copyupdate/'
+        handler = 'copyupdate'
         body = {'bucket': 'systemtest-textbucket'}
-        response = requests.post(url, json=body)
+        self.make_request(handler, body)
+
+    def make_request(self, handler, body):
+        url = f'http://{self._context.host_address}:{self._context.app_port}/{handler}/'
+        if handler == "deletefile":
+            response = requests.delete(url, json=body)
+        else:
+            response = requests.post(url, json=body)
         self._context.http_client_response = response
